refactor reviewDetails, remove useEffect, remove CommentForm.tsx and move the component into reviewDetails
