+++ frontend/src/components/CommentList.tsx
@@ -7,8 +7,7 @@ interface CommentListProps {
 
 export function CommentList({comments}: CommentListProps) {
     return <ul>
-        {comments
-            // todo: sort
+        {comments //todo: sort
             .map(c => <li
                 key={c.id}><CommentCard review={c}/>
             </li>)}
