+++ frontend/src/components/ReviewDetails.tsx
@@ -1,33 +1,51 @@
 import {Review} from "../lib/Review";
 import React, {useEffect, useState} from "react";
-import {CommentService, ReviewComment} from "../lib/ReviewComment";
+import {CommentRequest, CommentService, ReviewComment} from "../lib/ReviewComment";
 import {CommentList} from "./CommentList";
 import {ReviewCardWithoutLink} from "./ReviewCards";
 import "./ReviewDetails.css"
-import {CommentForm} from "./CommentForm";
+import {useParams} from "react-router-dom";
 
 interface ReviewDetailsProps {
     review: Review
 }
 
 export function ReviewDetails({review}: ReviewDetailsProps) {
+    const {courseId, authorUsername} = useParams();
     const [comments, setComments] = useState<ReviewComment[]>([]);
-    const [reloadCommentsFlag, setReloadCommentsFlag] = useState<boolean>(true);
+    const [newComment, setNewComment] = useState<string>("");
 
-    useEffect(() => {
-        console.log("in useEffect, reloadFlag is ", reloadCommentsFlag)
+    function reloadCourses() {
         CommentService.fetchCommentsByReview(review)
             .then(c => {
                 console.log(comments)
                 setComments(c)
             })
-        console.log(comments.at(comments.length-1))
-    }, [review, reloadCommentsFlag]);
+    }
 
+    function handleCommentSubmit() {
+        if (newComment == "") return; //todo: inform user comment can't be blank
+        const request: CommentRequest = {
+            text: newComment
+        }
+        CommentService.postComment(request, courseId!, authorUsername!)
+            .then(_ => {
+                reloadCourses();
+                setNewComment("")
+            })
+    }
 
+    reloadCourses();
     return <div>
         <ReviewCardWithoutLink review={review}/>
         <CommentList comments={comments}/>
-        <CommentForm reloadFlag={reloadCommentsFlag} reloadFlagSetter={setReloadCommentsFlag}/>
+        <div className="add-comment-container">
+            <textarea
+                placeholder="Twój komentarz"
+                onChange={e => setNewComment(e.target.value)}
+                value={newComment}
+            />
+            <button onClick={handleCommentSubmit}>Dodaj komentarz</button>
+        </div>
     </div>
 }
